--- frontend/src/utils/supabase_server.ts
@@ -1,16 +1,0 @@
-import { createServerClient, type CookieOptions } from '@supabase/ssr'
-import { cookies } from 'next/headers'
-
-const cookieStore = cookies()
-
-export const supabase = createServerClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    {
-        cookies: {
-        get(name: string) {
-            return cookieStore.get(name)?.value
-        },
-        },
-    }
-)
